requirements.txt
crewai
pyyaml
crewai-tools
python-dotenv
langchain-groq
langchain-google-genai
langchain-openai
